fix import syntax
